+++ src/test/java/io/github/acm19/aws/interceptorv5/test/Sample.java
@@ -30,6 +30,7 @@ import org.apache.hc.core5.http.message.BasicNameValuePair;
 import io.github.acm19.aws.interceptor.http.AwsRequestSigningApacheV5Interceptor;
 import software.amazon.awssdk.auth.credentials.DefaultCredentialsProvider;
 import software.amazon.awssdk.auth.signer.Aws4Signer;
+import software.amazon.awssdk.core.signer.Signer;
 import software.amazon.awssdk.regions.Region;
 
 class Sample {
@@ -38,25 +39,25 @@ class Sample {
     protected String endpoint;
     protected Region region;
     protected String service;
+    private final Signer signer;
 
-    Sample(final String service, final String[] args) throws ParseException {
+    Sample(String service, String[] args) throws ParseException {
+        this(service, args, Aws4Signer.create());
+    }
+
+    Sample(String service, String[] args, Signer signer) throws ParseException {
         this.service = service;
         parseOptions(args);
+        this.signer = signer;
     }
 
-    Sample(final String service, final String endpoint, final Region region) {
-        this.endpoint = endpoint;
-        this.region = region;
-        this.service = service;
-    }
-
-    public static void main(final String[] args) throws IOException, ParseException {
+    public static void main(String[] args) throws IOException, ParseException {
         Sample sampleClass = new Sample("", args);
         sampleClass.makeGetRequest();
         sampleClass.makePostRequest();
     }
 
-    private void parseOptions(final String[] args) throws ParseException {
+    private void parseOptions(String[] args) throws ParseException {
         Options options = new Options()
                 .addRequiredOption(null, "endpoint", true, "OpenSearch endpoint")
                 .addRequiredOption(null, "region", true, "AWS signing region");
@@ -98,7 +99,7 @@ class Sample {
         logRequest(httpPost);
     }
 
-    void logRequest(final HttpUriRequest request) throws IOException {
+    void logRequest(HttpUriRequest request) throws IOException {
         CloseableHttpClient httpClient = signingClient();
         try (CloseableHttpResponse response = httpClient.execute(request)) {
             switch (response.getCode()) {
@@ -114,7 +115,7 @@ class Sample {
     CloseableHttpClient signingClient() {
         HttpRequestInterceptor interceptor = new AwsRequestSigningApacheV5Interceptor(
                 service,
-                Aws4Signer.create(),
+                signer,
                 DefaultCredentialsProvider.create(),
                 region);