+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### 2.2.1 (Next)
 
+* [#90](https://github.com/acm19/aws-request-signing-apache-interceptor/pull/90): Add support for OpenSearch Serverless - [@acm19](https://github.com/acm19).
+
 ### 2.2.0 (2022/10/02)
 
 * [#80](https://github.com/acm19/aws-request-signing-apache-interceptor/pull/80): Add support for Apache client v5 - [@acm19](https://github.com/acm19).